Rename FlagsProvider-> Flags Client, FlagsClient -> FlagsClientMap and improve docs
