+++ features/dd-sdk-android-flags/api/dd-sdk-android-flags.api
@@ -22,23 +22,20 @@ public final class com/datadog/android/flags/FlagsConfiguration$Builder {
 	public final fun setEnableExposureLogging (Z)Lcom/datadog/android/flags/FlagsConfiguration$Builder;
 }
 
-public final class com/datadog/android/flags/featureflags/FlagsClient {
-	public static final field INSTANCE Lcom/datadog/android/flags/featureflags/FlagsClient;
-	public static final fun get ()Lcom/datadog/android/flags/featureflags/FlagsProvider;
-	public static final fun get (Lcom/datadog/android/api/SdkCore;)Lcom/datadog/android/flags/featureflags/FlagsProvider;
-	public static synthetic fun get$default (Lcom/datadog/android/api/SdkCore;ILjava/lang/Object;)Lcom/datadog/android/flags/featureflags/FlagsProvider;
-	public static final fun isRegistered ()Z
-	public static final fun isRegistered (Lcom/datadog/android/api/SdkCore;)Z
-	public static synthetic fun isRegistered$default (Lcom/datadog/android/api/SdkCore;ILjava/lang/Object;)Z
-}
-
-public abstract interface class com/datadog/android/flags/featureflags/FlagsProvider {
+public abstract interface class com/datadog/android/flags/featureflags/FlagsClient {
 	public abstract fun resolveBooleanValue (Ljava/lang/String;Z)Z
 	public abstract fun resolveIntValue (Ljava/lang/String;I)I
 	public abstract fun resolveNumberValue (Ljava/lang/String;Ljava/lang/Number;)Ljava/lang/Number;
 	public abstract fun resolveStringValue (Ljava/lang/String;Ljava/lang/String;)Ljava/lang/String;
 	public abstract fun resolveStructureValue (Ljava/lang/String;Lorg/json/JSONObject;)Lorg/json/JSONObject;
 	public abstract fun setContext (Lcom/datadog/android/flags/featureflags/model/EvaluationContext;)V
+}
+
+public final class com/datadog/android/flags/featureflags/FlagsClientMap {
+	public static final field INSTANCE Lcom/datadog/android/flags/featureflags/FlagsClientMap;
+	public static final fun instance ()Lcom/datadog/android/flags/featureflags/FlagsClient;
+	public static final fun instance (Lcom/datadog/android/api/SdkCore;)Lcom/datadog/android/flags/featureflags/FlagsClient;
+	public static synthetic fun instance$default (Lcom/datadog/android/api/SdkCore;ILjava/lang/Object;)Lcom/datadog/android/flags/featureflags/FlagsClient;
 }
 
 public final class com/datadog/android/flags/featureflags/model/EvaluationContext {
