+++ features/dd-sdk-android-flags/src/test/kotlin/com/datadog/android/flags/FlagsTest.kt
@@ -11,8 +11,8 @@ import com.datadog.android.api.InternalLogger
 import com.datadog.android.api.context.DatadogContext
 import com.datadog.android.core.InternalSdkCore
 import com.datadog.android.flags.Flags.FLAGS_EXECUTOR_NAME
-import com.datadog.android.flags.featureflags.FlagsClient
-import com.datadog.android.flags.featureflags.internal.NoOpFlagsProvider
+import com.datadog.android.flags.featureflags.FlagsClientMap
+import com.datadog.android.flags.featureflags.internal.NoOpFlagsClient
 import com.datadog.android.flags.internal.FlagsFeature
 import com.datadog.android.flags.internal.FlagsFeature.Companion.FLAGS_FEATURE_NAME
 import fr.xgouchet.elmyr.annotation.StringForgery
@@ -93,7 +93,7 @@ internal class FlagsTest {
     }
 
     @Test
-    fun `M register FlagsProvider W enable() { valid context }`() {
+    fun `M register FlagsClient W enable() { valid context }`() {
         // Given
         val fakeConfiguration = FlagsConfiguration.Builder()
             .setEnableExposureLogging(false)
@@ -103,12 +103,11 @@ internal class FlagsTest {
         Flags.enable(fakeConfiguration, mockSdkCore)
 
         // Then
-        assertThat(FlagsClient.isRegistered(mockSdkCore)).isTrue()
-        assertThat(FlagsClient.get(mockSdkCore)).isNotInstanceOf(NoOpFlagsProvider::class.java)
+        assertThat(FlagsClientMap.instance(mockSdkCore)).isNotInstanceOf(NoOpFlagsClient::class.java)
     }
 
     @Test
-    fun `M not register FlagsProvider W enable() { missing context }`() {
+    fun `M not register FlagsClient W enable() { missing context }`() {
         // Given
         val fakeConfiguration = FlagsConfiguration.Builder()
             .setEnableExposureLogging(false)
@@ -120,7 +119,7 @@ internal class FlagsTest {
         Flags.enable(fakeConfiguration, mockSdkCore)
 
         // Then
-        assertThat(FlagsClient.isRegistered(mockSdkCore)).isFalse()
+        assertThat(FlagsClientMap.instance(mockSdkCore)).isInstanceOf(NoOpFlagsClient::class.java)
     }
 
     @Test
