+++ features/dd-sdk-android-flags/src/test/kotlin/com/datadog/android/flags/featureflags/internal/DatadogFlagsClientTest.kt
@@ -30,11 +30,9 @@ import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.junit.jupiter.MockitoSettings
 import org.mockito.kotlin.any
 import org.mockito.kotlin.argumentCaptor
-import org.mockito.kotlin.capture
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.never
-import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import org.mockito.quality.Strictness
@@ -46,7 +44,7 @@ import java.util.concurrent.ExecutorService
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
 @ForgeConfiguration(ForgeConfigurator::class)
-internal class DatadogFlagsProviderTest {
+internal class DatadogFlagsClientTest {
 
     @Mock
     lateinit var mockExecutorService: ExecutorService
@@ -63,13 +61,13 @@ internal class DatadogFlagsProviderTest {
     @Mock
     lateinit var mockEvaluationsManager: EvaluationsManager
 
-    private lateinit var testedProvider: DatadogFlagsProvider
+    private lateinit var testedClient: DatadogFlagsClient
 
     @BeforeEach
     fun `set up`() {
         whenever(mockFeatureSdkCore.internalLogger) doReturn mockInternalLogger
 
-        testedProvider = DatadogFlagsProvider(
+        testedClient = DatadogFlagsClient(
             featureSdkCore = mockFeatureSdkCore,
             evaluationsManager = mockEvaluationsManager,
             flagsRepository = mockFlagsRepository
@@ -91,7 +89,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveBooleanValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveBooleanValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeFlagValue)
@@ -109,7 +107,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveBooleanValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveBooleanValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -123,7 +121,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn null
 
         // When
-        val result = testedProvider.resolveBooleanValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveBooleanValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -146,7 +144,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveStringValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStringValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeFlagValue)
@@ -160,7 +158,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn null
 
         // When
-        val result = testedProvider.resolveStringValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStringValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -183,7 +181,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveIntValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveIntValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeFlagValue)
@@ -201,7 +199,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveIntValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveIntValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -215,7 +213,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn null
 
         // When
-        val result = testedProvider.resolveIntValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveIntValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -238,7 +236,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveNumberValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveNumberValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeFlagValue)
@@ -252,7 +250,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn null
 
         // When
-        val result = testedProvider.resolveNumberValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveNumberValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -280,7 +278,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result.toString()).isEqualTo(fakeFlagValue.toString())
@@ -302,7 +300,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -330,7 +328,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -361,7 +359,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val result = testedProvider.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -385,7 +383,7 @@ internal class DatadogFlagsProviderTest {
         whenever(mockFlagsRepository.getPrecomputedFlag(fakeFlagKey)) doReturn null
 
         // When
-        val result = testedProvider.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
+        val result = testedClient.resolveStructureValue(fakeFlagKey, fakeDefaultValue)
 
         // Then
         assertThat(result).isEqualTo(fakeDefaultValue)
@@ -407,14 +405,14 @@ internal class DatadogFlagsProviderTest {
         whenever(customRepository.getPrecomputedFlag(fakeFlagKey)) doReturn fakeFlag
 
         // When
-        val provider = DatadogFlagsProvider(
+        val client = DatadogFlagsClient(
             featureSdkCore = mockFeatureSdkCore,
             evaluationsManager = mockEvaluationsManager,
             flagsRepository = customRepository
         )
 
         // Then
-        val result = provider.resolveStringValue(fakeFlagKey, "default")
+        val result = client.resolveStringValue(fakeFlagKey, "default")
         assertThat(result).isEqualTo(fakeFlagValue)
 
         verify(customRepository).getPrecomputedFlag(fakeFlagKey)
@@ -437,7 +435,7 @@ internal class DatadogFlagsProviderTest {
         val fakeContext = EvaluationContext(fakeTargetingKey, fakeAttributes)
 
         // When
-        testedProvider.setContext(fakeContext)
+        testedClient.setContext(fakeContext)
 
         // Then
         val contextCaptor = argumentCaptor<DatadogEvaluationContext>()
@@ -460,7 +458,7 @@ internal class DatadogFlagsProviderTest {
         val fakeAttributes = mapOf("test" to "value")
 
         // When
-        testedProvider.setContext(EvaluationContext(blankTargetingKey, fakeAttributes))
+        testedClient.setContext(EvaluationContext(blankTargetingKey, fakeAttributes))
 
         // Then
         verify(mockEvaluationsManager, never()).updateEvaluationsForContext(any())
@@ -473,7 +471,7 @@ internal class DatadogFlagsProviderTest {
         val fakeAttributes = mapOf("test" to "value")
 
         // When
-        testedProvider.setContext(EvaluationContext(blankTargetingKey, fakeAttributes))
+        testedClient.setContext(EvaluationContext(blankTargetingKey, fakeAttributes))
 
         // Then
         argumentCaptor<() -> String> {
@@ -504,7 +502,7 @@ internal class DatadogFlagsProviderTest {
         val fakeContext = EvaluationContext(fakeTargetingKey, fakeAttributes)
 
         // When
-        testedProvider.setContext(fakeContext)
+        testedClient.setContext(fakeContext)
 
         // Then
         // Verify that the evaluations manager was called to process the context
