+++ features/dd-sdk-android-flags/src/test/kotlin/com/datadog/android/flags/featureflags/internal/model/DatadogEvaluationContextTest.kt
@@ -244,7 +244,7 @@ internal class DatadogEvaluationContextTest {
         val context = DatadogEvaluationContext("user123", mapOf("plan" to "premium"))
 
         // When
-        val isValid = context?.isValid()
+        val isValid = context.isValid()
 
         // Then
         assertThat(context).isNotNull
